BUG: (issue #3) Fixed PlayerScreen based hooks overloading
